warning prevention
svn path=/trunk/FreeOrion/; revision=6524

+++ log4cpp/src/TimeStamp.cpp
@@ -35,7 +35,7 @@ namespace log4cpp {
 #ifdef LOG4CPP_HAVE_FTIME
         struct timeb tb;
         ::ftime(&tb);
-        _seconds = tb.time;
+        _seconds = static_cast<int>(tb.time);
         _microSeconds = 1000 * tb.millitm;
 #else
         _seconds = ::time(NULL);
